refactor: change the place of place of the three lines data, domain and ui to be in the scope of the home screen itself
