+++ lib/core/routing/generate_route.dart
@@ -3,7 +3,7 @@ import 'package:super_fitness/core/routing/defined_routes.dart';
 import 'package:super_fitness/modules/authentication/domain/entities/login/login_data_response_entity.dart';
 import 'package:super_fitness/modules/authentication/ui/login/login_screen.dart';
 import 'package:super_fitness/modules/authentication/ui/register/all_register_feature.dart';
-import 'package:super_fitness/modules/home/home_screen.dart';
+import 'package:super_fitness/modules/home/ui/home_screen.dart';
 import 'package:super_fitness/modules/onboarding/ui/screen/onboarding_screen.dart';
 
 abstract class GenerateRoute {
