+++ lib/modules/home/data/api/api_client_provider/home_api_client_provider.dart
@@ -1,0 +1,11 @@
+import 'package:dio/dio.dart';
+import 'package:injectable/injectable.dart';
+import 'package:super_fitness/modules/home/data/api/api_client/home_api_client.dart';
+
+@module
+abstract class HomeApiClientProvider {
+  @lazySingleton
+  HomeApiClient provideApiClient(Dio dio) {
+    return HomeApiClient(dio);
+  }
+}
