--- lib/modules/home/pages/workouts_page/data/api/api_client_provider/workout_api_client_provider.dart
@@ -1,12 +1,0 @@
-import 'package:dio/dio.dart';
-import 'package:injectable/injectable.dart';
-
-import '../api_client/workout_page_api_client.dart';
-
-@module
-abstract class WorkoutApiClientProvider {
-  @lazySingleton
-  WorkoutApiClient provideApiClient(Dio dio) {
-    return WorkoutApiClient(dio);
-  }
-}
