--- lib/modules/home/pages/workouts_page/domain/usecases/get_muscles_group_use_case.dart
@@ -1,15 +1,0 @@
-import 'package:injectable/injectable.dart';
-import 'package:super_fitness/modules/home/pages/workouts_page/domain/repo_contract/workout_repo_contract.dart';
-
-import '../../../../../../core/apis/api_result/api_result.dart';
-import '../entities/muscles_group_response_entity.dart';
-
-@injectable
-class GetMusclesGroupUseCase {
-  WorkoutRepo workoutRepo;
-
-  GetMusclesGroupUseCase(this.workoutRepo);
-
-  Future<ApiResult<List<MuscleGroupEntity>?>> execute() =>
-      workoutRepo.getMusclesGroup();
-}
