--- test/modules/home/pages/workouts_page/data/datasource_impl/workout_datasource_impl_test.mocks.dart
@@ -1,77 +1,0 @@
-// Mocks generated by Mockito 5.4.6 from annotations
-// in super_fitness/test/modules/home/pages/workouts_page/data/datasource_impl/workout_datasource_impl_test.dart.
-// Do not manually edit this file.
-
-// ignore_for_file: no_leading_underscores_for_library_prefixes
-import 'dart:async' as _i5;
-
-import 'package:mockito/mockito.dart' as _i1;
-import 'package:super_fitness/modules/home/pages/workouts_page/data/api/api_client/workout_page_api_client.dart'
-    as _i4;
-import 'package:super_fitness/modules/home/pages/workouts_page/data/models/get_all_muscles_group_response.dart'
-    as _i2;
-import 'package:super_fitness/modules/home/pages/workouts_page/data/models/muscle_group_workouts.dart'
-    as _i3;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: deprecated_member_use
-// ignore_for_file: deprecated_member_use_from_same_package
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: must_be_immutable
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-// ignore_for_file: subtype_of_sealed_class
-
-class _FakeGetMusclesGroupResponse_0 extends _i1.SmartFake
-    implements _i2.GetMusclesGroupResponse {
-  _FakeGetMusclesGroupResponse_0(Object parent, Invocation parentInvocation)
-    : super(parent, parentInvocation);
-}
-
-class _FakeMuscleGroupWorkoutsResponse_1 extends _i1.SmartFake
-    implements _i3.MuscleGroupWorkoutsResponse {
-  _FakeMuscleGroupWorkoutsResponse_1(Object parent, Invocation parentInvocation)
-    : super(parent, parentInvocation);
-}
-
-/// A class which mocks [WorkoutApiClient].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockWorkoutApiClient extends _i1.Mock implements _i4.WorkoutApiClient {
-  MockWorkoutApiClient() {
-    _i1.throwOnMissingStub(this);
-  }
-
-  @override
-  _i5.Future<_i2.GetMusclesGroupResponse> getMusclesGroup() =>
-      (super.noSuchMethod(
-            Invocation.method(#getMusclesGroup, []),
-            returnValue: _i5.Future<_i2.GetMusclesGroupResponse>.value(
-              _FakeGetMusclesGroupResponse_0(
-                this,
-                Invocation.method(#getMusclesGroup, []),
-              ),
-            ),
-          )
-          as _i5.Future<_i2.GetMusclesGroupResponse>);
-
-  @override
-  _i5.Future<_i3.MuscleGroupWorkoutsResponse> getMusclesGroupWorkouts(
-    String? id,
-  ) =>
-      (super.noSuchMethod(
-            Invocation.method(#getMusclesGroupWorkouts, [id]),
-            returnValue: _i5.Future<_i3.MuscleGroupWorkoutsResponse>.value(
-              _FakeMuscleGroupWorkoutsResponse_1(
-                this,
-                Invocation.method(#getMusclesGroupWorkouts, [id]),
-              ),
-            ),
-          )
-          as _i5.Future<_i3.MuscleGroupWorkoutsResponse>);
-}
