+++ test/modules/home/ui/pages/workouts_page/view_model/workouts_page_cubit_test.mocks.dart
@@ -1,0 +1,115 @@
+// Mocks generated by Mockito 5.4.6 from annotations
+// in super_fitness/test/modules/home/ui/pages/workouts_page/view_model/workouts_page_cubit_test.dart.
+// Do not manually edit this file.
+
+// ignore_for_file: no_leading_underscores_for_library_prefixes
+import 'dart:async' as _i4;
+
+import 'package:mockito/mockito.dart' as _i1;
+import 'package:mockito/src/dummies.dart' as _i7;
+import 'package:super_fitness/core/apis/api_result/api_result.dart' as _i5;
+import 'package:super_fitness/modules/home/domain/entities/workouts/muscle_group_workout_response_entity.dart'
+    as _i9;
+import 'package:super_fitness/modules/home/domain/entities/workouts/muscles_group_response_entity.dart'
+    as _i6;
+import 'package:super_fitness/modules/home/domain/repositories_contracts/workout_repo_contract.dart'
+    as _i2;
+import 'package:super_fitness/modules/home/domain/use_cases/workouts/get_muscle_group_workout_use_case.dart'
+    as _i8;
+import 'package:super_fitness/modules/home/domain/use_cases/workouts/get_muscles_group_use_case.dart'
+    as _i3;
+
+// ignore_for_file: type=lint
+// ignore_for_file: avoid_redundant_argument_values
+// ignore_for_file: avoid_setters_without_getters
+// ignore_for_file: comment_references
+// ignore_for_file: deprecated_member_use
+// ignore_for_file: deprecated_member_use_from_same_package
+// ignore_for_file: implementation_imports
+// ignore_for_file: invalid_use_of_visible_for_testing_member
+// ignore_for_file: must_be_immutable
+// ignore_for_file: prefer_const_constructors
+// ignore_for_file: unnecessary_parenthesis
+// ignore_for_file: camel_case_types
+// ignore_for_file: subtype_of_sealed_class
+
+class _FakeWorkoutRepo_0 extends _i1.SmartFake implements _i2.WorkoutRepo {
+  _FakeWorkoutRepo_0(Object parent, Invocation parentInvocation)
+    : super(parent, parentInvocation);
+}
+
+/// A class which mocks [GetMusclesGroupUseCase].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockGetMusclesGroupUseCase extends _i1.Mock
+    implements _i3.GetMusclesGroupUseCase {
+  MockGetMusclesGroupUseCase() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i2.WorkoutRepo get workoutRepo =>
+      (super.noSuchMethod(
+            Invocation.getter(#workoutRepo),
+            returnValue: _FakeWorkoutRepo_0(
+              this,
+              Invocation.getter(#workoutRepo),
+            ),
+          )
+          as _i2.WorkoutRepo);
+
+  @override
+  set workoutRepo(_i2.WorkoutRepo? _workoutRepo) => super.noSuchMethod(
+    Invocation.setter(#workoutRepo, _workoutRepo),
+    returnValueForMissingStub: null,
+  );
+
+  @override
+  _i4.Future<_i5.ApiResult<List<_i6.MuscleGroupEntity>?>> execute() =>
+      (super.noSuchMethod(
+            Invocation.method(#execute, []),
+            returnValue:
+                _i4.Future<_i5.ApiResult<List<_i6.MuscleGroupEntity>?>>.value(
+                  _i7.dummyValue<_i5.ApiResult<List<_i6.MuscleGroupEntity>?>>(
+                    this,
+                    Invocation.method(#execute, []),
+                  ),
+                ),
+          )
+          as _i4.Future<_i5.ApiResult<List<_i6.MuscleGroupEntity>?>>);
+}
+
+/// A class which mocks [GetMuscleGroupWorkoutUseCase].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockGetMuscleGroupWorkoutUseCase extends _i1.Mock
+    implements _i8.GetMuscleGroupWorkoutUseCase {
+  MockGetMuscleGroupWorkoutUseCase() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i2.WorkoutRepo get workoutRepo =>
+      (super.noSuchMethod(
+            Invocation.getter(#workoutRepo),
+            returnValue: _FakeWorkoutRepo_0(
+              this,
+              Invocation.getter(#workoutRepo),
+            ),
+          )
+          as _i2.WorkoutRepo);
+
+  @override
+  _i4.Future<_i5.ApiResult<List<_i9.MuscleEntity>?>> execute(String? id) =>
+      (super.noSuchMethod(
+            Invocation.method(#execute, [id]),
+            returnValue:
+                _i4.Future<_i5.ApiResult<List<_i9.MuscleEntity>?>>.value(
+                  _i7.dummyValue<_i5.ApiResult<List<_i9.MuscleEntity>?>>(
+                    this,
+                    Invocation.method(#execute, [id]),
+                  ),
+                ),
+          )
+          as _i4.Future<_i5.ApiResult<List<_i9.MuscleEntity>?>>);
+}
